cache_tables/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aws-psycopg2==1.3.8
    # via -r requirements.in
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via requests
idna==3.4
    # via requests
pyyaml==6.0.1
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
urllib3==1.26.17
    # via
    #   -r requirements.in
    #   requests
